# backend/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: ForteCommerce API
  description: |
    # ForteCommerce API

    Full-stack e-commerce platform with asynchronous checkout, campaign-driven promotions,
    and role-based access control.

    ## Authentication

    All endpoints (except `/auth/login`) require a JWT bearer token. Obtain a token by calling:

    ```
    POST /auth/login
    ```

    Then include the token in all subsequent requests:

    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## User Roles

    - **buyer**: Can browse products, submit checkouts, and manage their orders
    - **seller**: Can manage products and campaigns (in addition to buyer capabilities)

    ## Async Checkout Flow

    The ForteCommerce checkout is asynchronous to handle complex promotion calculations:

    1. **Submit checkout**: `POST /checkout` with a list of SKUs → returns `checkout_id` (202 Accepted)
    2. **Poll status**: `GET /checkout/{id}` until `status` = `completed` (with result) or `failed`
    3. **Confirm order**: `POST /checkout/{id}/confirm` → creates an order and returns order details

    Checkout sessions expire after 15 minutes if not confirmed. The backend processes promotions
    asynchronously and updates the checkout session with calculated totals and discounts.

    ## Error Responses

    All error responses use the standard `ErrorResponse` format with a unique error code and descriptive message:

    ```json
    {
      "success": false,
...
      "meta": {
...
      }
...

...

...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/auth/login` endpoint
  schemas:
...
    Meta:
      type: object
      description: Metadata attached to every response
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique request identifier for tracing
          example: "123e4567-e89b-12d3-a456-426614174000"
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when response was generated (RFC3339 format)
          example: "2024-01-15T10:30:00Z"
      required:
        - request_id
        - timestamp
    SuccessResponse:
      type: object
      description: Generic success response wrapper
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response payload (structure varies by endpoint)
        meta:
          $ref: '#/components/schemas/Meta'
      required:
        - success
        - meta
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error message
          example: "Invalid credentials"
        meta:
          $ref: '#/components/schemas/Meta'
      required:
        - success
        - message
        - meta

    # Authentication
...
    CheckoutItem:
      type: object
      description: Individual item in a checkout
      properties:
        sku:
          type: string
          description: Product SKU
          example: "SKU-001"
        name:
          type: string
          description: Product name
          example: "Widget A"
        qty:
          type: integer
          description: Quantity requested
          example: 2
        price:
          type: number
          format: double
          description: Unit price at checkout time
          example: 50.00
        total:
          type: number
          format: double
          description: qty × price
          example: 100.00
      required:
        - sku
        - name
        - qty
        - price
        - total
    AppliedPromotion:
      type: object
      description: Promotion applied to a checkout
      properties:
        name:
          type: string
          description: Campaign/promotion name
          example: "Summer Sale 2024"
        description:
          type: string
          description: Campaign description
          example: "Get 15% off on orders over $100"
        discount:
          type: number
          format: double
          description: Discount amount in currency units
          example: 15.00
      required:
        - name
        - description
        - discount

    # Orders
    Order:
      type: object
      description: Confirmed order from a completed checkout
      properties:
        id:
          type: string
          format: uuid
          description: Unique order identifier
          example: "550e8400-e29b-41d4-a716-446655440002"
        checkout_session_id:
          type: string
          format: uuid
          description: Reference to the checkout session that created this order
          example: "550e8400-e29b-41d4-a716-446655440001"
        status:
          type: string
          enum:
            - pending
            - paid
            - cancelled
          description: Current order status
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckoutItem'
          description: Items in this order
        promotions_applied:
          type: array
          items:
            $ref: '#/components/schemas/AppliedPromotion'
          description: Promotions applied when order was placed
        subtotal:
          type: number
          format: double
          description: Order subtotal before discounts
          example: 150.00
        total_discount:
          type: number
          format: double
          description: Total discount amount
          example: 25.00
        total:
          type: number
          format: double
          description: Final order total
          example: 125.00
        created_at:
          type: string
          format: date-time
          description: When the order was created
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - checkout_session_id
        - status
        - items
        - promotions_applied
        - subtotal
        - total_discount
        - total
        - created_at
        - updated_at

    # Campaigns
...
    PaymentStatusResponse:
      type: object
      description: Updated payment status
      properties:
        status:
          type: string
          enum: [paid, pending, cancelled]
          example: "paid"
      required:
        - status
...
security:
  - BearerAuth: []
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: List user's orders
      description: Retrieve all orders for the authenticated user
      operationId: listOrders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Retrieve details of a specific order
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Order ID
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/pay:
    post:
      tags:
        - Orders
      summary: Mark order as paid
      description: |
        Update an order status to paid.
        Typically called after payment processing is complete.
      operationId: payOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Order ID
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentStatusResponse'
        '400':
          description: Invalid order ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel an order
      description: |
        Cancel an order and release reserved inventory back to available stock.
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Order ID
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentStatusResponse'
        '400':
          description: Invalid order ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...